+++ docs/HISTORY.rst
@@ -4,9 +4,10 @@ Changelog
 1.0a3 (unreleased)
 ------------------
 
-- Use IEventRecurrence adapter to store attributes directly on the context.
-  This fixes that recurrence occurrences start and end dates were not indexed,
-  because the DateRecurringIndex had not access to the recurrence attribute.
+- Use custom behavior factory for IEventRecurrence to store attributes directly
+  on the context. This fixes that recurrence occurrences start and end dates
+  were not indexed, because the DateRecurringIndex had not access to the
+  recurrence attribute.
   [thet]
 
 - IRecurrence adapter returns now acquisition-wrapped occurrence