+++ docs/HISTORY.rst
@@ -4,6 +4,18 @@ Changelog
 1.0a3 (unreleased)
 ------------------
 
+- IEventSummary behavior added for body text on Dexterity event type,
+  as well as a SearchableText indexer adapter for the Dexterity event
+  type.
+  [seanupton]
+
+- Filter calendar portlet search URLs for each day to a whitelist of
+  event portal_type values.  Prevents non-event add-on types with
+  start/end fields from showing up in calendar, as defense against
+  unintended consequences (add-ons could explicitly override this 
+  template if they define additional Event types).
+  [seanupton]
+
 - API refactoring:
   * Move all generic interfaces to plone.event,
   * Extend IEventAccessor adapters to also be able to set attributes.